requirements.txt
psutil==7.0.0
pywebview==5.4
pyinstaller==6.12.0
